requirements.txt
streamlit
requests
pillow
fastapi
uvicorn
torch
torchvision
joblib
pandas
pillow
scikit-learn
python-multipart
